updated visits route, and changed migrations data type from integer to bigint

+++ db/migrate/20181018041421_create_user_challenges.rb
@@ -1,8 +1,8 @@
 class CreateUserChallenges < ActiveRecord::Migration[5.1]
   def change
     create_table :user_challenges do |t|
-      t.integer :user_id
-      t.integer :challenge_id
+      t.bigint :user_id
+      t.bigint :challenge_id
       t.boolean :emailed
       t.datetime :date_sent
       t.boolean :ran_tests
