+++ db/migrate/20181228200144_create_visitor_activities.rb
@@ -1,7 +1,7 @@
 class CreateVisitorActivities < ActiveRecord::Migration[5.1]
   def change
     create_table :visitor_activities do |t|
-      t.integer :visit_id
+      t.bigint :visit_id
       t.string :section_ids
       t.boolean :did_click_input
       t.text :email_input_value
